Star Wars API
Got all routes for the Star Wars API working. Had to delete the vehicles pug file and server code due to the the swapi routes being messed up on their end. Added images for the star wars API as well, and provided info on what numeric values to put in each input